chore: refactor Ref component to flatten React's tree
